Backend/requirements.txt
fastapi
uvicorn[standard]
python-chess
numpy
pydantic-settings
